+++ rust/cubesql/cubesql/src/compile/rewrite/language.rs
@@ -344,6 +344,7 @@ macro_rules! variant_field_struct {
                 AggregateFunction::ApproxMedian => "ApproxMedian",
                 AggregateFunction::BoolAnd => "BoolAnd",
                 AggregateFunction::BoolOr => "BoolOr",
+                AggregateFunction::PercentileCont => "PercentileCont",
             }
         );
     };
@@ -364,6 +365,7 @@ macro_rules! variant_field_struct {
                 BuiltinScalarFunction::Log => "Log",
                 BuiltinScalarFunction::Log10 => "Log10",
                 BuiltinScalarFunction::Log2 => "Log2",
+                BuiltinScalarFunction::Power => "Power",
                 BuiltinScalarFunction::Round => "Round",
                 BuiltinScalarFunction::Signum => "Signum",
                 BuiltinScalarFunction::Sin => "Sin",